add Printers refactor, UI improvements at Settings.
