+++ client/src/components/pages/Store/Settings/Printers/PrinterAddForm.jsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Button, Select, SelectItem, Switch } from "@heroui/react";
+import { Button, Card, CardBody, Select, SelectItem, Switch } from "@heroui/react";
 
 const PRINTER_TYPES = ["CUSTOMER"];
 
@@ -24,17 +24,17 @@ export function PrinterAddForm({
   t,
 }) {
   return (
-    <div className="rounded-lg border border-gray-200 bg-gray-50 p-4">
-      <div className="mb-4">
-        <h3 className="text-lg font-semibold text-green-900">
-          {t("cardPrinters.addTitle")}
-        </h3>
-        <p className="text-sm text-gray-600">
-          {t("cardPrinters.addDescription")}
-        </p>
-      </div>
+    <Card shadow="none" className="border border-gray-200 rounded-lg">
+      <CardBody className="flex flex-col gap-4 p-4">
+        <div>
+          <h3 className="text-sm sm:text-base font-semibold text-green-900">
+            {t("cardPrinters.addTitle")}
+          </h3>
+          <p className="text-xs sm:text-sm text-gray-600 mt-0.5">
+            {t("cardPrinters.addDescription")}
+          </p>
+        </div>
 
-      <div className="flex flex-col gap-4">
         <Select
           label={t("cardPrinters.typeLabel")}
           value={printerType}
@@ -81,32 +81,35 @@ export function PrinterAddForm({
             ))}
         </Select>
 
-        <div className="flex flex-wrap items-center justify-end gap-4">
-          <Switch
-            size="sm"
-            isSelected={isDefault}
-            onValueChange={onDefaultChange}
-          >
-            {t("cardPrinters.defaultLabel")}
-          </Switch>
+        <div className="flex items-center justify-between gap-3">
+          <div className="flex items-center gap-3 flex-wrap">
+            <Switch
+              size="sm"
+              isSelected={isDefault}
+              onValueChange={onDefaultChange}
+            >
+              <span className="text-xs sm:text-sm">{t("cardPrinters.defaultLabel")}</span>
+            </Switch>
 
-          <Switch
-            size="sm"
-            isSelected={enabled}
-            onValueChange={onEnabledChange}
-          >
-            {t("cardPrinters.enabledLabel")}
-          </Switch>
+            <Switch
+              size="sm"
+              isSelected={enabled}
+              onValueChange={onEnabledChange}
+            >
+              <span className="text-xs sm:text-sm">{t("cardPrinters.enabledLabel")}</span>
+            </Switch>
+          </div>
 
           <Button
             color="primary"
+            className="h-8 min-w-16 px-3 rounded-small sm:h-10 sm:min-w-20 sm:px-4 sm:rounded-medium bg-green-800 shrink-0"
             onPress={onSubmit}
             isDisabled={saving || loadingAvailable || !printerName || !templateName}
           >
             {t("cardPrinters.addButton")}
           </Button>
         </div>
-      </div>
-    </div>
+      </CardBody>
+    </Card>
   );
 }
