+++ client/src/components/pages/Store/Settings/Printers/index.js
@@ -1,0 +1,1 @@
+export * from "./Printers";
